requirements.txt
ccxt>=4.0.0
pandas>=2.0.0
fastapi
uvicorn
python-dotenv
requests
supabase
postgrest
scikit-learn
joblib>=1.3.0
numpy
pusher==3.3.2
redis>=5.0.0
xgboost>=2.0.0
PyYAML
openai
websockets
yfinance
lxml>=5.0.0
certifi
cryptography
